refactor: put scripts into modules for easier app-client imports
